% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA2-package.R
\docType{data}
\name{cma_impedance_functions}
\alias{cma_impedance_functions}
\title{Impedance functions for Census Metropolitan Areas and Census Agglomerations}
\format{
A data frame containing 675 rows and 12 variables; each row represents a unique function.
\describe{
\item{CMA}{Census Metropolitan Area or Census Agglomeration unique identifier.}
\item{Pr}{Province or Territory.}
\item{PwMode}{Transportation mode. Possible values: 'Bike', 'Car', 'Transit', and 'Walk.'}
\item{distribution}{Name of the function that best adjusted the duration values considering the specific location and transportation mode (can be "exponential", "norm", “lnorm”, “gamma” or “unif”)}.
\item{est_1}{Estimate of the first parameter for the function described in 'distribution.'}
\item{est_2}{Estimate of the second parameter for the function described in 'distribution.'}
\item{loglike}{Log-Likelihood obtained for the function 'distribution.'}
\item{AIC}{Akaike information criterion obtained for function 'distribution.'}
\item{BIC}{Bayesian information criterion obtained for function 'distribution.'}
}
}
\source{
"2021 Census of Population, informations about the census is available \href{https://www12.statcan.gc.ca/census-recensement/2021/dp-pd/prof/index.cfm?Lang=E}{here} accessed December 19th 2024.
}
\usage{
data(cma_impedance_functions)
}
\description{
This object contains impedance functions for Census Metropolitan Areas
and Census Agglomerations (CMA/CAs), for job destinations,
considering different transportation modes ('Bike', 'Car', 'Transit', and 'Walk').
We estimated impedance functions  using the fitdistrplus package, using the 'PwDUR'
(Commuting duration) of the 2021 Census of Population.
The best-fit function was defined by selecting the lowest Akaike information criterion.
}
\keyword{agglomerations}
\keyword{area}
\keyword{ca}
\keyword{census}
\keyword{cma}
\keyword{decay}
\keyword{distance}
\keyword{functions}
\keyword{impedance}
\keyword{metropolitan}
