% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA2-package.R
\docType{data}
\name{cma_max_travel_times}
\alias{cma_max_travel_times}
\title{Maximum travel times in the Census Metropolitan Area.}
\format{
A simple feature class (sf) polygon object containing 596 rows and 3 variables; each row represents a combination of CMA and transportation mode.
\describe{
\item{CMA}{Uniquely identifies a CMA.}
\item{PwMode}{Transportation mode. Possible values: 'Bike', 'Car', 'Transit', and 'Walk.'}
\item{Max_time_value}{Maximum travel time for the CMA using the mentioned transportation mode.}
}
}
\source{
"\href{https://www12.statcan.gc.ca/census-recensement/2021/dp-pd/prof/index.cfm?Lang=E}{Census of Population 2021} accessed September 11, 2024.
}
\usage{
data(cma_max_travel_times)
}
\description{
This object contains the maximum travel times for commuting to work for all Census Metropolitan Area (CMA) in Canada,
obtained from the Census of Population (2021). The methodology to obtain this data set is described in the
R markdown 'making-data-v2.Rmd'.
}
\keyword{area}
\keyword{census}
\keyword{in}
\keyword{maximum}
\keyword{metropolitan}
\keyword{the}
\keyword{time}
\keyword{travel}
