% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA2-package.R
\docType{package}
\name{CommuteCA2-package}
\alias{CommuteCA2}
\alias{CommuteCA2-package}
\title{CommuteCA2: Standardized Methods for Accessibility Analysis in Canadian Research}
\description{
The main objective of the *CommuteCA* package is to develop standardized methods for accessibility analysis in research, especially for studies using Statistics Canada surveys. Among the available surveys, we focused our efforts on the 2021 Census of Population, which contain valuable variables for transportation research. This package was created with support from the office of the Research Data Center at McMaster University, the Sherman Centre for Digital Scholarship and Mobilizing Justice.
}
\author{
\strong{Maintainer}: Bruno Dos Santos \email{dossanb@mcmaster.ca} (\href{https://orcid.org/0000-0002-6748-2038}{ORCID})

Authors:
\itemize{
  \item Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})
}

}
\keyword{internal}
