% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommuteCA2-package.R
\docType{data}
\name{pcd_max_travel_times}
\alias{pcd_max_travel_times}
\title{Maximum travel times in the Census Division.}
\format{
A simple feature class (sf) polygon object containing 1,083 rows and 3 variables; each row represents a combination of CD and transportation mode.
\describe{
\item{PCD}{Uniquely identifies a census division (composed of the 2-digit province or territory unique identifier followed by the 2-digit census division code).}
\item{PwMode}{Transportation mode. Possible values: 'Bike', 'Car', 'Transit', and 'Walk.'}
\item{Max_time_value}{Maximum travel time for the CD using the mentioned transportation mode.}
}
}
\source{
"\href{https://www12.statcan.gc.ca/census-recensement/2021/dp-pd/prof/index.cfm?Lang=E}{Census of Population 2021} accessed September 11, 2024.
}
\usage{
data(pcd_max_travel_times)
}
\description{
This object contains the maximum travel times for commuting to work for all Census Divisions (CD) in Canada,
obtained from the Census of Population (2021). The methodology to obtain this data set is described in the
R markdown 'making-data-v2.Rmd'.
}
\keyword{census}
\keyword{division}
\keyword{in}
\keyword{maximum}
\keyword{the}
\keyword{time}
\keyword{travel}
